Implement q_ascend and q_descend function
Add q_ascend and q_descend function that remove every node which has a
node with a strictly greater/less value anywhere to the right side of
it.The function check from backward if current value is greater/less
than maximum/minimum, if so, change the maximum/minimum. If not delete
current node and move to the next node.

Change-Id: I62a17aced94af458beb72ec26cf16c63d18fb43f