Implement q_merge function
Add q_merge function that merge all the queues into one sorted queue.The
function count the needed merge times,then iterate through every queue.
For every iteration, input the first and second entry of the queue into
merge_queue function.The auxiliary function take the node of two entry
and compare it for every loop.Put the smaller one into the temp queue.
Finally, spice the rest of the remain queue to the tail of the new
queue.

Change-Id: Ibec82e4ddd550fb74a94f1b99e3c3fe1c79b140a